Quản lý size và color thêm sửa xóa có xóa mềm và khôi phục xóa

+++ database/migrations/2025_03_01_130834_add_is_hidden_to_comments_table.php
@@ -6,20 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
-    public function up()
+    public function up(): void
     {
         Schema::table('comments', function (Blueprint $table) {
-            $table->boolean('is_hidden')->default(false);
+            if (!Schema::hasColumn('comments', 'is_hidden')) {
+                $table->boolean('is_hidden')->default(0);
+            }
         });
     }
-    
-    public function down()
+
+    public function down(): void
     {
         Schema::table('comments', function (Blueprint $table) {
-            $table->dropColumn('is_hidden');
+            if (Schema::hasColumn('comments', 'is_hidden')) {
+                $table->dropColumn('is_hidden');
+            }
         });
     }
 };
